requirements.txt
# requirements.txt for Flask Azure Maps App
Flask
requests
python-dotenv
gunicorn
